requirements.txt
pyyaml
requests
gelfclient
influxdb